Rename Gamemodes
TDM is actually Elimination
TDMZ is actually Domination
DOM is actually TDM
I'm horrible at naming things
