+++ Gamemode Mods/Stable/StarCore SUGMA Gamemodes/Data/Scripts/SUGMA/GameModes/TeamDeathMatch/TeamDeathmatchGamemode.cs
@@ -1,39 +1,29 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Sandbox.Game.Entities;
 using Sandbox.ModAPI;
 using SC.SUGMA.API;
 using SC.SUGMA.GameState;
 using SC.SUGMA.Utilities;
 using VRage.Game.ModAPI;
 
-namespace SC.SUGMA.GameModes.TeamDeathMatch
+namespace SC.SUGMA.GameModes.TeamDeathmatch
 {
-    /// <summary>
-    ///     Each faction starts with 1200 seconds. Deaths remove a fraction of seconds.
-    /// </summary>
     internal class TeamDeathmatchGamemode : GamemodeBase
     {
-        public static double MatchDuration = 20;
+        public static double MatchDuration = 15;
+        public static int VictoryPoints = -1;
+        public static double RespawnTimeSeconds = 10;
 
-        /// <summary>
-        ///     Lists currently tracked factions. Mapped to grid count.
-        /// </summary>
-        public readonly Dictionary<IMyFaction, int> TrackedFactions = new Dictionary<IMyFaction, int>();
-
-        private bool _remOuter = false, _remMiddle = false, _remInner = false;
-
-        internal IMyFaction _winningFaction;
-
-
-        public PointTracker PointTracker;
-
-        internal ShareTrackApi ShareTrackApi => SUGMA_SessionComponent.I.ShareTrackApi;
-        internal MatchTimer _matchTimer => SUGMA_SessionComponent.I.GetComponent<MatchTimer>("MatchTimer");
         public override string ReadableName { get; internal set; } = "Team Deathmatch";
+        public override string Description { get; internal set; } = "Teams fight to for who can get the most kills within a time limit.";
 
-        public override string Description { get; internal set; } =
-            "Factions fight against eachother until tickets run out. Kill enemy players to remove tickets.";
+        internal static ShareTrackApi ShareTrackApi => SUGMA_SessionComponent.I.ShareTrackApi;
+
+        internal PointTracker PointTracker;
+        internal RespawnManager RespawnManager;
+        internal MatchTimer MatchTimer => SUGMA_SessionComponent.I.GetComponent<MatchTimer>("MatchTimer");
+
+        internal IMyFaction WinningFaction;
 
         public TeamDeathmatchGamemode()
         {
@@ -42,9 +32,23 @@ namespace SC.SUGMA.GameModes.TeamDeathMatch
                     time => double.TryParse(time, out MatchDuration),
                     "t",
                     "match-time",
-                    "Match time, in minutes.")
+                    "Match time, in minutes."),
+                new ArgumentParser.ArgumentDefinition(
+                    time => int.TryParse(time, out VictoryPoints),
+                    "k",
+                    "victory-kills",
+                    "Number of kills to end the match."),
+                new ArgumentParser.ArgumentDefinition(
+                    time => double.TryParse(time, out RespawnTimeSeconds),
+                    "r",
+                    "respawn-time",
+                    "Respawn time in seconds.")
             );
         }
+
+        internal HashSet<IMyFaction> TrackedFactions = new HashSet<IMyFaction>();
+
+
 
         public override void Close()
         {
@@ -53,130 +57,66 @@ namespace SC.SUGMA.GameModes.TeamDeathMatch
 
         public override void UpdateActive()
         {
-            if (PointTracker == null || _matchTimer == null ||
-                TrackedFactions == null) // ten billion nullchecks of aristeas
-                return;
-
-            foreach (var factionKvp in TrackedFactions.Keys.ToArray())
-                if (CalculateFactionPoints(factionKvp) <= 0)
-                    // Stop updating if the game just ended
-                    if (OnFactionKilled(factionKvp))
-                        return;
-
-            //double matchRatio = 1 - currentPoints / basePoints;
-            //
-            //if (matchRatio <= 0.76) // If within 12 seconds of 15:00 matchtime (assuming 20:00 total)
-            //{
-            //    if (matchRatio >= 0.75)
-            //        MyAPIGateway.Utilities.ShowNotification($"Removing outer cover in T-{basePoints*0.25 - currentPoints:N1}", 1000/60, "Red");
-            //    else if (!_remOuter)
-            //    {
-            //        RemoveBlockers(9750);
-            //        _remOuter = true;
-            //    }
-            //}
-            //
-            //if (matchRatio <= 0.51) // If within 12 seconds of 10:00 matchtime
-            //{
-            //    if (matchRatio >= 0.5)
-            //        MyAPIGateway.Utilities.ShowNotification($"Removing middle cover in T-{basePoints*0.5 - currentPoints:N1}", 1000/60, "Red");
-            //    else if (!_remMiddle)
-            //    {
-            //        RemoveBlockers(5750);
-            //        _remMiddle = true;
-            //    }
-            //}
-            //
-            //if (matchRatio <= 0.26) // If within 12 seconds of 5:00 matchtime
-            //{
-            //    if (matchRatio >= 0.25)
-            //        MyAPIGateway.Utilities.ShowNotification($"Removing inner cover in T-{basePoints*0.75 - currentPoints:N1}", 1000/60, "Red");
-            //    else if (!_remInner)
-            //    {
-            //        RemoveBlockers(1750);
-            //        _remInner = true;
-            //    }
-            //}
+            //if (PointTracker == null || MatchTimer == null ||
+            //    TrackedFactions == null) // ten billion nullchecks of aristeas
+            //    return;
+            if (MatchTimer.IsMatchEnded)
+                StopRound();
         }
 
-        private void RemoveBlockers(float maxDistanceFromCenter)
+        public override void StartRound(string[] arguments)
         {
-            var covers = new List<IMyCubeGrid>();
-            MyAPIGateway.Entities.GetEntities(null, ent =>
-            {
-                var grid = ent as IMyCubeGrid;
-                if (grid == null || grid.DisplayName != "#EntityCover" || ((MyCubeGrid)grid).BlocksCount > 1 ||
-                    grid.GetPosition().Length() <= maxDistanceFromCenter)
-                    return false;
-
-                covers.Add(grid);
-
-                return false;
-            });
-            foreach (var grid in covers)
-                grid.Close();
-        }
-
-        public override void StartRound(string[] arguments = null)
-        {
-            PointTracker = new PointTracker(3, 0);
-            SUGMA_SessionComponent.I.UnregisterComponent("TDMPointTracker");
-            if (!MyAPIGateway.Utilities.IsDedicated)
-                SUGMA_SessionComponent.I.UnregisterComponent("tdmHud");
-
             foreach (var grid in ShareTrackApi.GetTrackedGrids())
             {
-                var faction = PlayerTracker.I.GetGridFaction(grid);
-                if (faction == null || !ShareTrackApi.IsGridAlive(grid))
-                    continue;
-
-                if (!TrackedFactions.ContainsKey(faction))
-                    TrackedFactions.Add(faction, 1);
-                else
-                    TrackedFactions[faction]++;
+                if (grid.GetFaction() != null)
+                    TrackedFactions.Add(grid.GetFaction());
             }
 
-            if (TrackedFactions.Count <= 1)
+            if (TrackedFactions.Count < 2)
             {
                 MyAPIGateway.Utilities.ShowNotification("There aren't any combatants, idiot!", 10000, "Red");
                 StopRound();
                 return;
             }
 
+            base.StartRound(arguments);
+
+            PointTracker = new PointTracker(0, VictoryPoints);
+            RespawnManager = new RespawnManager(RespawnTimeSeconds);
             SUGMA_SessionComponent.I.RegisterComponent("TDMPointTracker", PointTracker);
+            SUGMA_SessionComponent.I.RegisterComponent("TDMRespawnManager", RespawnManager);
 
             ShareTrackApi.RegisterOnAliveChanged(OnAliveChanged);
-            ShareTrackApi.RegisterOnTrack(OnGridTrackChanged);
-
-            var factionNames = new List<string>();
-            foreach (var factionKvp in TrackedFactions)
+            PointTracker.OnFactionWin += faction =>
             {
-                PointTracker.SetFactionPoints(factionKvp.Key, factionKvp.Value);
-                factionNames.Add($"|{factionKvp.Key.Tag}|");
-                foreach (var faction in TrackedFactions.Keys)
+                WinningFaction = faction;
+                StopRound();
+            };
+            
+            var factionNames = new List<string>();
+            foreach (var faction in TrackedFactions)
+            {
+                factionNames.Add($"|{faction.Tag}|");
+                foreach (var otherFaction in TrackedFactions)
                 {
-                    if (faction == factionKvp.Key)
+                    if (otherFaction == faction)
                         continue;
 
-                    MyAPIGateway.Session.Factions.DeclareWar(factionKvp.Key.FactionId, faction.FactionId);
+                    MyAPIGateway.Session.Factions.DeclareWar(faction.FactionId, otherFaction.FactionId);
                     //MyAPIGateway.Utilities.ShowMessage("TDM", $"Declared war between {factionKvp.Key.Name} and {faction.Name}");
                 }
             }
 
-            base.StartRound(arguments);
             MyAPIGateway.Utilities.ShowNotification("Combatants: " + string.Join(" vs ", factionNames), 10000, "Red");
-            _matchTimer.Start(MatchDuration);
+            MatchTimer.Start(MatchDuration);
 
             if (!MyAPIGateway.Utilities.IsDedicated)
                 SUGMA_SessionComponent.I.RegisterComponent("tdmHud", new TeamDeathmatchHud(this));
 
-            string trackedGrids = "";
-            foreach (var faction in TrackedFactions)
-                trackedGrids += $"\n-   {faction.Key.Name}: {faction.Value} | {CalculateFactionPoints(faction.Key)}pts";
-
             Log.Info("Started a TDM match." +
-                     $"\n- Combatants: {string.Join(" vs ", factionNames)}" +
-                     "\n- Tracked grids:" + trackedGrids);
+                     $"\nDuration: {MatchDuration}m" + 
+                     $"\nVictory Points: {VictoryPoints}" + 
+                     $"\n- Combatants: {string.Join(" vs ", factionNames)}");
         }
 
         public override void StopRound()
@@ -190,7 +130,7 @@ namespace SC.SUGMA.GameModes.TeamDeathMatch
                     long factionId;
                     long.TryParse(arg.Remove(0, 3), out factionId);
 
-                    _winningFaction = MyAPIGateway.Session.Factions.TryGetFactionById(factionId);
+                    WinningFaction = MyAPIGateway.Session.Factions.TryGetFactionById(factionId);
                     setWinnerFromArgs = true;
                     break;
                 }
@@ -201,104 +141,62 @@ namespace SC.SUGMA.GameModes.TeamDeathMatch
                 if (!MyAPIGateway.Session.IsServer)
                     return;
 
-                Arguments = Arguments.Concat(new[] { $"win{_winningFaction?.FactionId ?? -1}" }).ToArray();
+                if (WinningFaction == null)
+                    WinningFaction = PointTracker.FactionPoints.MaxBy(f => f.Value).Key;
+
+                Arguments = Arguments.Concat(new[] { $"win{WinningFaction?.FactionId ?? -1}" }).ToArray();
             }
 
-            SUGMA_SessionComponent.I.GetComponent<TeamDeathmatchHud>("tdmHud")?.MatchEnded(_winningFaction);
+            SUGMA_SessionComponent.I.GetComponent<TeamDeathmatchHud>("tdmHud")?.MatchEnded(WinningFaction);
 
             foreach (var factionKvp in TrackedFactions)
-            foreach (var faction in TrackedFactions.Keys)
             {
-                if (faction == factionKvp.Key)
-                    continue;
+                foreach (var faction in TrackedFactions)
+                {
+                    if (faction == factionKvp)
+                        continue;
 
-                MyAPIGateway.Session.Factions.SendPeaceRequest(factionKvp.Key.FactionId, faction.FactionId);
-                MyAPIGateway.Session.Factions.AcceptPeace(faction.FactionId, factionKvp.Key.FactionId);
+                    MyAPIGateway.Session.Factions.SendPeaceRequest(factionKvp.FactionId, faction.FactionId);
+                    MyAPIGateway.Session.Factions.AcceptPeace(faction.FactionId, factionKvp.FactionId);
+                }
             }
+            
 
-            _matchTimer?.Stop();
+            MatchTimer?.Stop();
             SUGMA_SessionComponent.I.UnregisterComponent("TDMPointTracker");
+            SUGMA_SessionComponent.I.UnregisterComponent("TDMRespawnManager");
             ShareTrackApi.UnregisterOnAliveChanged(OnAliveChanged);
-            ShareTrackApi.UnregisterOnTrack(OnGridTrackChanged);
 
             base.StopRound();
-            _winningFaction = null;
+            WinningFaction = null;
             TrackedFactions.Clear();
             PointTracker = null;
+            RespawnManager = null;
         }
 
         internal override void DisplayWinMessage()
         {
-            if (_winningFaction == null)
+            if (WinningFaction == null)
             {
                 MyAPIGateway.Utilities.ShowNotification("YOU ARE ALL LOSERS.", 10000, "Red");
                 return;
             }
 
-            MyAPIGateway.Utilities.ShowNotification($"A WINNER IS [{_winningFaction?.Name}]!", 10000);
+            MyAPIGateway.Utilities.ShowNotification($"A WINNER IS [{WinningFaction?.Name}]!", 10000);
         }
 
-
-        public virtual int CalculateFactionPoints(IMyFaction faction)
+        private void OnAliveChanged(IMyCubeGrid grid, bool isAlive)
         {
-            if (!TrackedFactions.ContainsKey(faction))
-                return -1;
-
-            return (int)(_matchTimer.MatchDurationMinutes * 60 *
-                         (PointTracker.GetFactionPoints(faction) / (float)TrackedFactions[faction]) -
-                         _matchTimer.CurrentMatchTime.TotalSeconds);
-        }
-
-        internal virtual void OnAliveChanged(IMyCubeGrid grid, bool isAlive)
-        {
-            Log.Info("GridAliveSet: " + grid.DisplayName + " -> " + isAlive);
-
-            var gridFaction = PlayerTracker.I.GetGridFaction(grid);
-            if (gridFaction == null)
+            if (isAlive)
                 return;
 
-            PointTracker.AddFactionPoints(gridFaction, isAlive ? 1 : -1);
-            if (isAlive) MyAPIGateway.Utilities.ShowMessage("TDM", $"[{grid.DisplayName}] has returned to life!");
-            //int newPoints = PointTracker.GetFactionPoints(gridFaction);
-            //if (newPoints > TrackedFactions.GetValueOrDefault(gridFaction))
-            //    TrackedFactions[gridFaction] = newPoints; // TODO the UI will break a little bit if this gets called.
-        }
+            var gridFaction = grid.GetFaction();
 
-        internal virtual void OnGridTrackChanged(IMyCubeGrid grid, bool isTracked)
-        {
-            Log.Info("GridTrackSet: " + grid.DisplayName + " -> " + isTracked);
-
-            if (ShareTrackApi.IsGridAlive(grid))
+            foreach (var faction in TrackedFactions)
             {
-                var gridFaction = PlayerTracker.I.GetGridFaction(grid);
-                if (gridFaction == null)
-                    return;
-
-                PointTracker.AddFactionPoints(gridFaction, -1);
+                if (faction != gridFaction)
+                    PointTracker.AddFactionPoints(faction, 1);
             }
-        }
-
-        internal virtual bool OnFactionKilled(IMyFaction faction)
-        {
-            TrackedFactions.Remove(faction);
-            MyAPIGateway.Utilities.ShowNotification($"|{faction.Tag}| IS OUT!", 10000, "Red");
-            Log.Info($"|{faction.Tag}| was killed!");
-
-            if (TrackedFactions.Count == 1)
-            {
-                _winningFaction = TrackedFactions.Keys.First();
-                StopRound();
-                return true;
-            }
-
-            if (TrackedFactions.Count == 0)
-            {
-                _winningFaction = null;
-                StopRound();
-                return true;
-            }
-
-            return false;
         }
     }
 }
